+++ packages/config-tools/tsup.config.ts
@@ -10,11 +10,8 @@ export default defineConfig({
   clean: true,
   dts: true,
   sourcemap: false,
-  bundle: true,
-  tsconfig: "./tsconfig.json"
-  // dts: {
-  //   resolve: true,
-  //   // build types for `src/index.ts` only
-  //   entry: "./src/index.ts"
-  // }
+  tsconfig: "./tsconfig.json",
+  shims: true,
+  noExternal: ["chalk"],
+  skipNodeModulesBundle: true
 });