+++ packages/esbuild/src/plugins/node-protocol.ts
@@ -1,10 +1,14 @@
-import type { Plugin } from "esbuild";
+import type * as esbuild from "esbuild";
+import { ESBuildOptions, ESBuildResolvedOptions } from "../types";
 
 /**
  * The node: protocol was added to require in Node v14.18.0
  * https://nodejs.org/api/esm.html#node-imports
  */
-export const nodeProtocolPlugin = (): Plugin => {
+export const nodeProtocolPlugin = (
+  options: ESBuildOptions,
+  resolvedOptions: ESBuildResolvedOptions
+): esbuild.Plugin => {
   const nodeProtocol = "node:";
 
   return {