+++ packages/unbuild/src/index.ts
@@ -26,4 +26,6 @@
 
 export * from "./build";
 export * from "./clean";
+export * from "./config";
 export * from "./types";
+export * from "./utilities";